add fragranceFinder
