+++ mobile/src/screens/FragranceFinder/FragranceFinderSurvey.js
@@ -1,0 +1,42 @@
+
+import React from 'react';
+import {View, Text, Button} from 'react-native';
+import SkeletonPlaceholder from "react-native-skeleton-placeholder";
+
+function FragranceFinderSurvey({ route, navigation }) {
+  /* 2. Get the param */
+  /* https://reactnavigation.org/docs/6.x/params */
+  const { itemId, otherParam } = route.params;
+
+  return (
+    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+      <SkeletonPlaceholder>
+        <View style={{ flexDirection: "row", alignItems: "center" }}>
+          <View style={{ width: 60, height: 60, borderRadius: 50 }} />
+          <View style={{ marginLeft: 20 }}>
+            <View style={{ width: 120, height: 20, borderRadius: 4 }} />
+            <View
+              style={{ marginTop: 6, width: 80, height: 20, borderRadius: 4 }}
+            />
+          </View>
+        </View>
+      </SkeletonPlaceholder>      
+      <Text>Details Screen</Text>
+      <Text>itemId: {JSON.stringify(itemId)}</Text>
+      <Text>otherParam: {JSON.stringify(otherParam)}</Text>
+      <Button
+        title="Go to Details... again"
+        onPress={() =>
+          navigation.push('FragranceFinderSurvey', {
+            itemId: Math.floor(Math.random() * 100),
+            otherParam:'???????'
+          })
+        }
+      />
+      <Button title="Go to Home" onPress={() => navigation.navigate('Home')} />
+      <Button title="Go back" onPress={() => navigation.goBack()} />
+    </View>
+  );
+}
+
+export default FragranceFinderSurvey;
